Moved assignShifts's controller, type, and validation to feature/shift/assign
