+++ src/features/shift/assign/index/type.ts
@@ -1,0 +1,13 @@
+import type { AssignShift } from "@prisma/client";
+import type { ZodIssue } from "zod";
+
+export interface AssigShiftResponse {
+	ok: true;
+	assignShift: AssignShift;
+}
+
+export interface UpsertAssignShiftValidationErrorResponse {
+	ok: false;
+	message: string;
+	errors: ZodIssue[];
+}
